requirements.txt
streamlit>=1.20
pandas>=1.5
numpy>=1.23
pdfplumber>=0.7.6
python-docx>=0.8.11
reportlab>=3.7.0
google-genai>=0.2.0
google.generativeai
python-dotenv
docx
PyPDF2
openai

